+++ modules/global/src/de/diedavids/cuba/userinbox/entity/Message.java
@@ -130,4 +130,8 @@ public class Message extends StandardEntity {
     private UserSessionSource getUserSessionSource() {
         return AppBeans.get(UserSessionSource.class);
     }
+
+    public void toggleRead() {
+        setRead(!getRead());
+    }
 }